requirements.txt
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
pefile==2019.4.18
pipreqs==0.4.11
psycopg2==2.9.3
requests==2.26.0
torch==1.9.0
typing-extensions==3.10.0.0
urllib3==1.26.7
whitenoise==5.3.0
